Unit tests Iteration#19

+++ src/test/java/dev/defaultybuf/feathercore/modules/common/utils/TempFile.java
@@ -6,13 +6,14 @@
  *
  * @file TempFile.java
  * @author Alexandru Delegeanu
- * @version 0.1
+ * @version 0.2
  * @description Helper class to delete given file when no longer needed
  */
 package dev.defaultybuf.feathercore.modules.common.utils;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import org.apache.commons.io.FileUtils;
 
@@ -26,7 +27,9 @@ public class TempFile implements AutoCloseable {
     @Override
     public void close() {
         try {
-            FileUtils.delete(path.toFile());
+            if (Files.exists(path)) {
+                FileUtils.delete(path.toFile());
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
